requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.2.0
numpy>=2.0.0
python-dotenv>=1.0.0
click>=8.1.7
schedule>=1.2.0
json5>=0.9.14
lxml>=4.9.3
selenium>=4.15.2
webdriver-manager>=4.0.1
scikit-learn>=1.3.2
flask>=3.0.0